% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_dir.R
\name{neon_db_dir}
\alias{neon_db_dir}
\title{Default directory for persistent NEON database}
\usage{
neon_db_dir()
}
\value{
the active \code{neonstore} directory.
}
\description{
Use \code{neon_db_dir()} to view or access the currently active database
directory. By default, this uses the appropriate application directory
for your operating system, see \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}.
This location can be overridden by setting
the environmental variable \code{NEONSTORE_DB}.
}
\examples{

neon_db_dir()

## Override with an environmental variable:
Sys.setenv(NEONSTORE_DB = tempdir())
neon_db_dir()
## Unset
Sys.unsetenv("NEONSTORE_DB")

}
