% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_db.R
\name{neon_disconnect}
\alias{neon_disconnect}
\title{Disconnect from the neon database}
\usage{
neon_disconnect(db = neon_db())
}
\arguments{
\item{db}{link to an existing database connection}
}
\description{
Disconnect from the neon database
}
