% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{neon_download}
\alias{neon_download}
\title{product <- "DP1.10003.001"
".\emph{basic.}\link{.}zip"}
\usage{
neon_download(
  product,
  dest = neon_dir(),
  file_regex = "[.]csv",
  api = "https://data.neonscience.org/api/v0"
)
}
\description{
product <- "DP1.10003.001"
".\emph{basic.}\link{.}zip"
}
