% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_export_db.R
\name{neon_export_db}
\alias{neon_export_db}
\title{Export NEON database to parquet}
\usage{
neon_export_db(dir = file.path(neon_dir(), "parquet"), db = neon_db())
}
\arguments{
\item{dir}{directory to which parquet export is written.}

\item{db}{Connection to your local NEON database}
}
\description{
Export your current database. This can be important to (1)
archive and share your database files with another user or machine,
(2) expose your database using an S3 bucket using neon_remote_db(),
(3) assist in upgrading your duckdb version.
}
