% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_export_db.R
\name{neon_import_db}
\alias{neon_import_db}
\title{Import a NEON database exported from neon_export_db()}
\usage{
neon_import_db(
  dir = file.path(neon_dir(), "parquet"),
  db = neon_db(read_only = FALSE)
)
}
\arguments{
\item{dir}{directory to which parquet export is written.}

\item{db}{Connection to your local NEON database}
}
\description{
Import a NEON database exported from neon_export_db()
}
