% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{neon_index}
\alias{neon_index}
\title{Show information about all files downloaded to the local store}
\usage{
neon_index(product = NULL, table = NULL, hash = "md5", dir = neon_dir())
}
\arguments{
\item{product}{Include only files matching this NEON productCode (optional)}

\item{table}{Include only files matching this table name (or regex pattern).
(optional).}

\item{hash}{name of a hashing algorithm to check file integrity. Can be
md5, sha1, or sha256 currently; or set to \link{NULL} to skip hash computation.}

\item{dir}{Location where files should be downloaded. By default will
use the appropriate applications directory for your system
(see \link[rappdirs:user_data_dir]{rappdirs::user_data_dir}).  This default also be configured by
setting the environmental variable \code{NEONSTORE_HOME}, see \link{Sys.setenv} or
\link{Renviron}.}
}
\description{
NEON products consist of several individual components, which are in turn
broken up by site and sampling month. By storing these individual files,
neonstore enables more reproducible workflows that can be traced back to
original, unaltered input data.  These atomized files can be quickly and easily
combined into unified tables, see \link{neon_read}.
}
\details{
File names include metadata such as the file productCode,
table name, site, and sampling month, as well as timestamp of creation.
\code{neon_index()} parses this metadata from the file name string and returns
the information in a convenient table, along with a path to each file.
}
\examples{

\dontshow{
# Hide setting tempfile, since a user would specify a persistent location
 Sys.setenv("NEONSTORE_HOME"=tempfile())
}

neon_index()

## Just bird survey product
neon_index("DP1.10003.001")

\dontshow{
# tidy
 Sys.unsetenv("NEONSTORE_HOME")
}

}
