% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_pane.R
\name{neon_pane}
\alias{neon_pane}
\title{Open NEON database connection pane in RStudio}
\usage{
neon_pane()
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver"))) neon_pane()
}
