% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{neon_products}
\alias{neon_products}
\title{Table of all NEON Data Products}
\usage{
neon_products(
  fields = c("productCode", "productName", "productDescription", "productStatus",
    "themes", "keywords", "productCategory", "productAbstract",
    "productDesignDescription", "productRemarks", "productSensor",
    "productPublicationFormatType", "productHasExpanded", "productBasicDescription",
    "productExpandedDescription"),
  api = "https://data.neonscience.org/api/v0",
  .token = Sys.getenv("NEON_TOKEN")
)
}
\arguments{
\item{fields}{a list of fields (columns) to include.  Default includes
most common columns, set to NULL to display all columns
(including list-columns).}

\item{api}{the URL to the NEON API, leave as default.}

\item{.token}{an authentication token from NEON. A token is not
required but will allow access to a higher number of requests before
rate limiting applies, see
\url{https://data.neonscience.org/data-api/rate-limiting/#api-tokens}.
Note that once files are downloaded once, \code{neonstore} provides persistent
access to them without further interaction required with the API.}
}
\description{
Return a table of all NEON Data Products, including product descriptions
and the productCode needed for \link{neon_download}.
}
\examples{
\donttest{

products <- neon_products()

# Or search for a keyword
i <- grepl("bird", products$keywords)
products[i, c("productCode", "productName")]

}
}
\seealso{
\link{neon_download}
}
