% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_store.R
\name{neon_registry}
\alias{neon_registry}
\title{neon registry}
\usage{
neon_registry(path = neon_registry_default())
}
\arguments{
\item{path}{specify a path for the registry, or use the default}
}
\value{
Will create an empty cv file at the registry if no exists
}
\description{
neon registry
}
