% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_remote.R
\name{neon_remote}
\alias{neon_remote}
\title{neon_remote

select a table from the remote connection}
\usage{
neon_remote(table = "", product = "", type = "", db = neon_remote_db())
}
\arguments{
\item{table}{table name (pattern match regex)}

\item{product}{product code}

\item{type}{basic or extended (if necessary to distinguish)}

\item{db}{a \link{neon_remote_db} connection.  If not provided, one will be created,
but it is faster to pass this on for re-use in multiple \code{neon_remote} calls.}
}
\value{
a arrow::FileSystemDataset object, or a named list of such
objects if multiple matches are found.  This table is not downloaded
but remains on the remote storage location, but can be filtered
with dplyr functions like filter and select, and can also be
grouped and summarised, all without ever downloading the whole table.
Use \code{\link[dplyr:compute]{dplyr::collect()}} to download the (possibly filtered) table into
and pull into memory.
}
\description{
neon_remote

select a table from the remote connection
}
