% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_export_db.R
\name{neon_share_db}
\alias{neon_share_db}
\title{sync local parquet export to an S3 database}
\usage{
neon_share_db(s3, dir = file.path(neon_dir(), "parquet"))
}
\arguments{
\item{s3}{an \verb{[arrow::s3_bucket()]} connection or other valid arrow path}

\item{dir}{directory to which parquet export is written.}
}
\description{
sync local parquet export to an S3 database
}
