% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_sites.R
\name{neon_sites}
\alias{neon_sites}
\title{Table of all NEON sites}
\usage{
neon_sites(
  api = "https://data.neonscience.org/api/v0",
  .token = Sys.getenv("NEON_TOKEN")
)
}
\arguments{
\item{api}{the URL to the NEON API, leave as default.}

\item{.token}{an authentication token from NEON. A token is not
required but will allow access to a higher number of requests before
rate limiting applies, see
\url{https://data.neonscience.org/data-api/rate-limiting/#api-tokens}.
Note that once files are downloaded once, \code{neonstore} provides persistent
access to them without further interaction required with the API.}
}
\description{
Returns a table of all NEON sites by making a single API call
to the \verb{/sites} endpoint.
}
