% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_store.R
\name{neon_stack}
\alias{neon_stack}
\title{Stack zip files into combined tables using the local store}
\usage{
neon_stack(
  dpID = NA,
  dir = contentid::content_dir(),
  registry = neon_registry(),
  workdir = fs::path_temp("temporary"),
  nCores = parallel::detectCores()
)
}
\arguments{
\item{dpID}{Data product ID of product to stack. Not needed; defaults to NA, included for back compatibility}

\item{dir}{location to use for the content store}

\item{registry}{location to write the local neon data registry}

\item{workdir}{Location to use for temporary storage, will be purged after use}

\item{nCores}{The number of cores to parallelize the stacking procedure. To automatically use the maximum number of cores on your machine we suggest setting nCores=parallel::detectCores(). By default it is set to a single core.}
}
\value{
ids and names of stacked tables
}
\description{
Stack zip files into combined tables using the local store
}
\examples{
\donttest{
neon_store("DP1.10003.001")
neon_stack("DP1.10003.001")
}
}
