% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{neon_store}
\alias{neon_store}
\title{Show information about all files downloaded to the local store}
\usage{
neon_store(table = NULL, dir = neon_dir())
}
\arguments{
\item{table}{a table name (or regex pattern) to filter on.}

\item{dir}{Location where files should be downloaded. By default will
use the appropriate applications directory for your system
(see \link[rappdirs:user_data_dir]{rappdirs::user_data_dir}).  This default also be configured by
setting the environmental variable \code{NEONSTORE_HOME}, see \link{Sys.setenv} or
\link{Renviron}.}
}
\description{
Show information about all files downloaded to the local store
}
