% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_export_db.R
\name{neon_sync_db}
\alias{neon_sync_db}
\title{sync local parquet export to an S3 database}
\usage{
neon_sync_db(to, from = local_bucket())
}
\arguments{
\item{to}{an \verb{[arrow::SubTreeFileSystem]}, such as a remote connection to
an S3 bucket from \verb{[arrow::s3_bucket()]}.}

\item{from}{another \verb{[arrow::SubTreeFileSystem]}, such as local path.
By default, this is the same default path used by \verb{[neon_import_db()]}
and \verb{[neon_export_db()]}}
}
\description{
sync local parquet export to an S3 database
}
