% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_table.R
\name{neon_table}
\alias{neon_table}
\title{Return a neon table from the database}
\usage{
neon_table(table, site = NA, dir = neon_dir())
}
\arguments{
\item{table}{the name of a downloaded NEON table in the store,
see \link{neon_index}}

\item{site}{4-letter site code(s) to filter on. Leave as \code{NA} to search all.}

\item{dir}{Location where files should be downloaded. By default will
use the appropriate applications directory for your system
(see \link[rappdirs:user_data_dir]{rappdirs::user_data_dir}).  This default also be configured by
setting the environmental variable \code{NEONSTORE_HOME}, see \link{Sys.setenv} or
\link{Renviron}.}
}
\description{
Return a neon table from the database
}
\details{
We cannot filter on start_date or end_date since these
come only from the filename metadata and are only added
to instrument tables, not observation tables etc.
}
