% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_export_db.R
\name{standardize_export_names}
\alias{standardize_export_names}
\title{standardize export names}
\usage{
standardize_export_names(dir = file.path(neon_dir(), "parquet"))
}
\arguments{
\item{dir}{directory to which parquet export is written.}
}
\description{
standardize export names
}
\details{
DUCKDB clobbers database filenames to avoid potentially incompatible characters.
This is pretty unnecessary, so we can restore the original table names for
use with S3-based remote access which assumes parquet files map to the
desired table names (i.e. including product numbers.)

However, note that \verb{[neon_import_db()]} uses native duckdb functions
that assume the original mangled names.
}
