% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_db.R
\name{neon_disconnect}
\alias{neon_disconnect}
\title{Disconnect from the neon database}
\usage{
neon_disconnect(dir = neon_dir())
}
\arguments{
\item{dir}{Location where files should be downloaded. By default will
use the appropriate applications directory for your system
(see \link[rappdirs:user_data_dir]{rappdirs::user_data_dir}).  This default also be configured by
setting the environmental variable \code{NEONSTORE_HOME}, see \link{Sys.setenv} or
\link{Renviron}.}
}
\description{
Disconnect from the neon database
}
