% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_remote.R
\name{neon_remote_db}
\alias{neon_remote_db}
\title{Establish a remote database connection using \code{arrow}}
\usage{
neon_remote_db(
  s3 = arrow::s3_bucket("neon", endpoint_override = "data.ecoforecast.org")
)
}
\arguments{
\item{s3}{an \code{arrow::s3_bucket()} connection}

\item{host}{An S3-compliant host address}

\item{bucket}{bucket name}

\item{...}{additional arguments to \code{arrow::s3_bucket()}}
}
\description{
Provides a remote connection
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

db <- neon_remote_db()
\dontshow{\}) # examplesIf}
}
