% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_remote.R
\name{neon_remote_db}
\alias{neon_remote_db}
\title{Establish a remote database connection using \code{arrow}}
\usage{
neon_remote_db(
  bucket = arrow::s3_bucket("neon4cast-targets/neon", endpoint_override =
    "data.ecoforecast.org")
)
}
\arguments{
\item{bucket}{an \verb{[arrow::s3_bucket]} connection or other
\link[arrow:FileSystem]{arrow::SubTreeFileSystem} object.}
}
\description{
Establish a remote database connection using \code{arrow}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

db <- neon_remote_db()
\dontshow{\}) # examplesIf}
}
