% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_store.R
\name{neon_store}
\alias{neon_store}
\title{download and store a data product}
\usage{
neon_store(
  dpID,
  site = "all",
  startdate = NA,
  enddate = NA,
  package = "basic",
  avg = "all",
  check.size = TRUE,
  savepath = NA,
  load = F,
  registry = neon_registry()
)
}
\arguments{
\item{dpID}{The identifier of the NEON data product to pull, in the form DPL.PRNUM.REV, e.g. DP1.10023.001}

\item{site}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to all.}

\item{startdate}{Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to NA.}

\item{enddate}{Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to NA.}

\item{package}{Either 'basic' or 'expanded', indicating which data package to download. Defaults to basic.}

\item{avg}{Either the string 'all', or the averaging interval to download, in minutes. Only applicable to sensor (IS) data. Defaults to 'all'.}

\item{check.size}{T or F, should the user approve the total file size before downloading? Defaults to T. When working in batch mode, or other non-interactive workflow, use check.size=F.}

\item{savepath}{The location to save the output files to}

\item{load}{T or F, are files saved locally or loaded directly? Used silently with loadByProduct(), do not set manually.}

\item{registry}{location to write the local neon data registry}
}
\value{
path to the neon data registry csv file (invisibly)
}
\description{
download and store a data product
}
\details{
This is a very simple wrapper around zipsByProduct.  zipsByProduct downloads
the requested data as a set of .zip files to local disk.  We jsut want to know
if those files are the same each time, or have changed.
}
\examples{
\donttest{
neon_store("DP1.10003.001")
 }


}
