% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_export_db.R
\name{neon_sync_db}
\alias{neon_sync_db}
\title{sync local parquet export to an S3 database}
\usage{
neon_sync_db(s3, dir = file.path(neon_dir(), "parquet"))
}
\arguments{
\item{s3}{an \verb{[arrow::SubTreeFileSystem]}, such as a remote connection to
an S3 bucket from \verb{[arrow::s3_bucket()]}.}

\item{dir}{directory to which parquet export is written.}
}
\description{
sync local parquet export to an S3 database
}
\details{
Remote files are named according to the table name (including
product id, not according to the 'sanitized' file name duckdb uses
when generating exports.)
}
