% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neon_table.R
\name{neon_table}
\alias{neon_table}
\title{Return a neon table from the database}
\usage{
neon_table(table, site = NA, db = neon_db())
}
\arguments{
\item{table}{the name of a downloaded NEON table in the store,
see \link{neon_index}}

\item{site}{4-letter site code(s) to filter on. Leave as \code{NA} to search all.}

\item{db}{a connection to the database, see \verb{[neon_db()]}.}
}
\description{
Return a neon table from the database
}
\details{
We cannot filter on start_date or end_date since these
come only from the filename metadata and are only added
to instrument tables, not observation tables etc.
}
